+++ src/types.ts
@@ -36,7 +36,6 @@ export type {
   CollectionDescription,
   CollectionEncryption,
   CollectionEncryptionEpoch,
-  CollectionEncryptionEpochsMac,
   CollectionEncryptionRecipient,
   PolicyDocument,
   LinkSet,