+++ test/node/edv-doc-cipher.test.ts
@@ -31,7 +31,6 @@ import {
   isEncryptedEnvelope
 } from '../../src/edv/index.js'
 import { mintEpoch, wrapEpochSecret } from '../../src/edv/epochCrypto.js'
-import { computeEpochsMac } from '../../src/edv/epochMac.js'
 import type {
   CollectionEncryption,
   CollectionEncryptionRecipient
@@ -84,10 +83,6 @@ async function epochDescriptorFor(
     epochs: [{ id: epochId, recipients }],
     currentEpoch: epochId
   }
-  encryption.epochsMac = await computeEpochsMac({
-    descriptor: encryption,
-    epochSecret: secret
-  })
   return encryption
 }